requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.10
# by the following command:
#
#    pip-compile pyproject.toml
#
anyio==4.4.0
    # via watchfiles
argon2-cffi==23.1.0
    # via django-coding-challenge (pyproject.toml)
argon2-cffi-bindings==21.2.0
    # via argon2-cffi
asgiref==3.8.1
    # via
    #   django
    #   django-cors-headers
async-timeout==4.0.3
    # via redis
cffi==1.16.0
    # via argon2-cffi-bindings
click==8.1.7
    # via uvicorn
colorama==0.4.6
    # via
    #   click
    #   uvicorn
django==4.2.14
    # via
    #   django-coding-challenge (pyproject.toml)
    #   django-cors-headers
    #   django-redis
    #   djangorestframework
django-cors-headers==4.4.0
    # via django-coding-challenge (pyproject.toml)
django-environ==0.11.2
    # via django-coding-challenge (pyproject.toml)
django-redis==5.4.0
    # via django-coding-challenge (pyproject.toml)
djangorestframework==3.15.2
    # via django-coding-challenge (pyproject.toml)
exceptiongroup==1.2.2
    # via anyio
h11==0.14.0
    # via uvicorn
hiredis==3.0.0
    # via redis
httptools==0.6.1
    # via uvicorn
idna==3.7
    # via anyio
jinja2==3.1.4
    # via django-coding-challenge (pyproject.toml)
markupsafe==2.1.5
    # via jinja2
pycparser==2.22
    # via cffi
python-dotenv==1.0.1
    # via uvicorn
pyyaml==6.0.1
    # via uvicorn
redis[hiredis]==5.0.8
    # via
    #   django-coding-challenge (pyproject.toml)
    #   django-redis
sniffio==1.3.1
    # via anyio
sqlparse==0.5.1
    # via django
typing-extensions==4.12.2
    # via
    #   anyio
    #   asgiref
    #   uvicorn
tzdata==2024.1
    # via django
uvicorn[standard]==0.29.0
    # via django-coding-challenge (pyproject.toml)
watchfiles==0.22.0
    # via uvicorn
websockets==12.0
    # via uvicorn
